cambio int por double

+++ Ejemplo36.java
@@ -5,15 +5,15 @@ public class Ejemplo36 {
     public static void main(String[] args){
 
         //Declaro variables
-        int[] numero = {1,2,3,4,5,6,7,8,9,10};
-        int mayor = 0;
+        double[] numero = {1,2,3,4,5,6,7,8,9,10};
+        double mayor = 0;
         Scanner reader = new Scanner (System.in);
 
         //Empieza el programa
-        for (int valor : numero) {
+        for (double valor : numero) {
 
             System.out.println("Dame diez numeros: (llevas "+ (valor-1) +")");
-            valor = reader.nextInt();
+            valor = reader.nextDouble();
                 if (mayor == 0){ mayor = valor;
                 }
